[build] Make the build system cleaner and more modular
Each extension can now be enabled or disabled throught the appropriate
configure flag : --enable-extensionname or --disable-extensionname.
If no flag is given for an extension, it will be built if the
dependencies are present, and disabled if they are not.
If an extension is explicitly enabled and the dependencies are not
satisfied, configure stops with an error.

Each extension has its own macro in a specific file, which makes for a
cleaner configure.ac.