[AppIndicator] Fixing toggle state issue when the window was launched from its starter
When Banshee would be launched via its starter in e.g. the menu or a dock when it was
already running, the visibility of its main window would be toggled. However, the
toggle state of the menu item in the AppIndicator menu would not be updated accordingly.
Now fixed.
(Bug found by Chow Loong Jin)