[SoundMenu] Make it also conflict with the NotificationArea extension
As AppIndicator, having SoundMenu and NotificationArea loaded at
the same time leads to undesirable side-effects. SoundMenu now detects
if it was enabled after NotificationArea and disables it if so. It will
also disable itself if NotificationArea is enabled last.
This is of course a blatant copy of the previous commit by Iain Lane.