[build] Bring gnome-doc-utils autofoo into the tree
gnome-doc-utils' build is a nightmare. I try to make this a little saner
and safer by not running gnome-doc-prepare in autogen. Doing this the
way we were doing it means it's possible for a tarball to be released
without the necessary m4 and .make includes.

It's very tricky and gross to /optionally/ support gnome-doc-utils in
the build. Because of this, both the required .m4 and .make files are
included explicitly in the tree now, which means they must be updated as
well, when fixes are introduced upstream. Tough luck.

Also, I reintroduce the fact that ENABLE_SK is only defined when
GNOME_DOC_INIT is called. When this is called, there's a mandatory
pkg-config requirement for gnome-doc-utils. Therefore, we only call
GNOME_DOC_INIT if we actually want to enable gnome-doc-utils support. If
the support is disabled, we still need to ensure ENABLE_SK is defined,
so this is now defined to false before any possible call to
GNOME_DOC_INIT, which can override ENABLE_SK to whatever it pleases.

Gross.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>