[AlarmClock] Rework synchronization and timeout for the alarm thread
Use an AutoResetEvent instead of interrrupting or aborting the thread,
and trigger the alarm with a GLib.Timeout. This should prevent an
infinite loop in some cases, where the thread would not go back to sleep
after being interrupted for some reason.

An insomniac thread in an alarm clock, go figure !