[Karaoke] Create new Karaoke extension to filter the singers voice
This is the first basic version of the working Karaoke extension
that filters the singers voice from normal music that is played in
the Banshee player. Karaoke mode is currently switched on and off by
enabling/disabling the extension.

Uses Banshee.Streamrecorder.Gst namespace. Some functionality has
been added there.