[Karaoke] Added on/off toggle button and config dialog and removed false dependency
Added a configuration dialog to set the effect parameters. Also added
a toggle button in the header toolbar to switch the Karaoke effect on
and off without disabling the extension. Removed addin.xml dependency
to Streamrecorder as that required the Streamrecorder extension to be
enabled. The library is required, but the extension does not need to
be enabled for Karaoke to work.