Streamrecorder: rebuild Recorder logic for fixing bgo#642933
Adding the recorder at delayed initialization once and add an output-
selector into the recorder pipeline that is used to switch from a
fakesink to a filesink. That way asynchronous pad blocking while
handling connecting to and from audiotee becomes obsolete and the
recorder can work more seemsless and the stream mentioned in the bug
can be played and recorded, as well as other streams. Also cleaned up
EOS event as it duplicates other events.