Fixed the launchpad bug #748120 which takes care of differenciating between ScheduledActivity and UserActivity
The URL of the bug is https://bugs.launchpad.net/zeitgeist-dataproviders/+bug/748120
Title:
Enabling Zeitgeist data-logger in Banshee clutters the lenses
Description:
I want my recently played songs to show up in, for example, the Synapse launcher, I
enabled the Zeitgeist data logger extension in Banshee. However, due to this all my
recently played songs show up under "recent files" in the files lens. Consequently,
it becomes practically unusable.

Mikkel wrote the root of the issue:
This is actually a very interesting observation that will affect any "generic recent
files list" based on Zeitgeist.

My guess is that the Banshee history just describes all music playing activities as
an event of typezg: UserActivity an a subject that is a nfo:FileDataObject and has
interpretation nfo:Audio or something.

What I believe is the "correct" behaviour for the music player is to:

 - When the user manually selects a song log as described above
 - When the music player automatically skips to the next song the manifestation
type of the event should be zg:ScheduledActivity[1]

With this logging pattern the "generic recent files list" will only contain the songs
you manually selected for playback which I think is "correct" FWIW :-) (and will make
the Unity lenses work without modification (== less work for me yay!))

[1]: The docs for zg:ScheduledActivity in fact reads: "An event that was directly triggered
by some user initiated sequence of actions. For example a music player automatically
changing to the next song in a playlist".