Drop liblirc_client.so moduleref
We can't resolve the SONAME of liblirc_client.so during build time, so we can't
write a proper dllmap for liblirc_client.so. We can't just moduleref on
liblirc_client.so either, since this is a build time symlink that is only
present in -dev packages (on distros that do the -dev split).

This commit solves the above issue by adding some wrapper functions in the glue
library which just forward the arguments on.

This fixes LP #773686