Drop crlf normalizing for sln files
According to gitattributes(5), setting the crlf attribute causes git to
normalize the line endings while checking in, and convert them back to CRLF when
checking out.

The implications of this are that:
 - The *.sln files stored as git blobs in the index and commit trees *should* have
   LF eols.

 - The checked out *.sln files *should* have CRLF eols.

What actually happened is that the *.sln files keep getting checked in with CRLF
eols (probably due to a defunct git implementation somewhere. git on Windows,
perhaps?). This causes a newly cloned git tree to appear dirty with newer git
installations, with `git reset --hard' not cleaning it up.

Seeing as we always want the *.sln files to appear with CRLF eols, this commit
removes the crlf attribute, causing the files to have CRLF eols both in the
index, and when checked out.