ClutterFlow: Don't update the texture size while we're running
Updating the texture size on the fly was not working well and causing
lots of complications, so remove support for that.

Add a little note in the preferences page to explain that you need to
restart Banshee for changes to the dimensions preferences to take
effect. Ideally we should just choose a reasonable value and not show
those preferences at all.