LastfmFingerprint: Add dllmap for liblastfmfpbridge
Although mono tries many funky permutations of the dll name, it is only
documented to try searching for "lib${moduleref}.so"[1].

Hence, the moduleref of liblastfmfpbridge is insufficient, and a dllmap is
needed to map this appropriately. Dropping the "lib" from the moduleref would
not work either, as Windows does not prepend "lib" to the library name while
when loading modulerefs.

[1] http://www.mono-project.com/Interop_with_Native_Libraries