Fanart: avoid NullReferenceException when there is no logo
When there is no logo for a certain artist, Fanart response is simply
a string that contains the word "null". When this happened, it wouldn't
be parsed properly as JSON, obviously, which meant that the call to
method JsonObject.FromString() was returning null, which later was passed
to the constructor of CorrectResults, which would generate a NRE, which
would be caught by a try-catch which would call the constructor of
IncorrectResults class, which would receive the exception.

Seems simple enough to just convert all this cruft into code that doesn't
catch exceptions, but which also prints a warning when the response is
neither JSON nor "null".