Let MonoDevelop 4.0 re-arrange the .sln and .csproj files
This is just the result of opening the solution properties without
changing anything, to let MonoDevelop 4.0 re-arrange the .sln and
.csproj files as it wants. This should prevent noise in later commits
that contain real changes.

This also add a proper GUID for the CueSheets project.

Some projects are temporarily downgraded from the 4.0 framework to 3.5,
for consistency, but everybody will be bumped shortly.