Switch csproj files to Mono/.NET 4.0 target framework
This also upgrades the solution file to the Visual Studio 2010 format,
and updates Hyena to the latest git master, that now also targets the
4.0 framework.

This only impacts the build when using xbuild or building from the IDE.
The usual 'make' build is not affected.

Also add the SoundCloud csproj in the solution, so that it gets updated.