SongKick: add sketch
From time to time, SongKick should search for concerts
(of the user's favorite artists) which take place
in the user's city, and throw a notification if there is any.

This commit is just a sketch.
Here you can see what is going on when SongKick finds a gig
which take place near user:
https://mega.co.nz/#!7FcnXBaY!viWf5V1qkGJk_iW3F-kSdFuGe5P5OdQNG3UkHcZ-UgI