Dap.Bluetooth: Add UI 'No Adapters' display
The main source page previously assumed there would be adapters to
control so the "Power" and "Discovery" buttons were still sensitive,
although obviously not operational, in the instances where no adapter was
installed in the user's system.

This patch adds an informative message when no adapters are present and
disables the buttons described above in those cases too.