Dap.Bluetooth: Use Localizer, tweak TrackInfo load
After some confusion, it appears that
AddinManager.CurrentLocalizer.GetString is preferred over
Catalog.GetString for localised messages. This patch adds a couple of
functions to cover the correct method and make use of them in the new
source.

Slight performance tweak in that the major file list is not reversed
when loading information, but yet to extract
size/created/accessed/modified properties from the dictionary - a new
structural type will be used for this, which we'll cache.

Also, I was mistaken on the purpose of the TrackEqualHandler. I was
hoping it would allow me to match up remote tracks with ones in the
user's main library. This patch provides a more sane implementation
based on its apparent purpose; determining if the currently playing
track is the one it is checking against.