Dap.Bluetooth: Implements add and delete track ops
Prior to this patch we had a read-only source. Along with track
add/delete operations this patch extracts the ListFolder output into the
more friendly RemoteNode record along with some improvements to the
functions for inferring metadata.

In addition, an extra function is provided for obtaining the appropriate
extension of a file based on its mime type.

Finally, updates formatting of match expressions in MimeApi to follow on
a new line, this appears best for layout.