Dap.Bluetooth: Correct added tracks' pseudo-URI
Correct an error in the path list construction for added tracks,
causing added tracks to be 'misplaced' if attempting to delete them
during the same session.

This patch also makes a DatabaseTrackInfo copy after I realised that the
modify/save routine I was doing was causing the track to be removed from
its parent playlist.

Sync now works, in that already added tracks are not re-synchronised
during the same session. Matching remote track information with
previously transferred tracks will allow us to prevent the resync by
ensuring the MetadataHash property is available when we need it.