Dap.Bluetooth: Fuzzy Lookup
In order to support sync, banshee needs to have the MetadataHash
available, calculated from certain fields including the Genre and Year.
Since the bluetooth FTP connection provides only path information it is
not possible to gather enough to calculate the hash, this is made more
difficult by escaped special characters in file names.

This patch attempts a fuzzy lookup on files located within a heirarchy
and, it is hoped, will still function even on difficult names; since
banshee uses the sqlite wildcard character for escaping.