Dap.Bluetooth: Images reworked, drops RSSI text
Images specified with the IconName attribute appear to be 22px, although
this may vary between systems? Hopefully not, the two icons not found in
the system set are now 22px square.

Patch also adds an attribution for paired-black.png. The unicode image
is public domain.

Finally, the RSSI (signal strength) was mostly only useful for
debugging, and is only available when the adapter is discovering, and
then only if the device is visible (in spite of settings such as
"visible only to paired devices" on some bluetooth equipment) so it's
not really useful especially since the UI now reacts to the Connected
property of the device, bolding the Alias.