Dap.Bluetooth: Corrects Stale Properties
The property manager requires signaling when an object is removed from
the bus, as (obviously, when thought about) there is no property
invalidation signal coming from it when it dies.

In addition, a check is added for cases where a device properties
changed signal occurs when the device is yet to be mapped to a widget.