Dap.Bluetooth: Include gnome interop module
In order to control the software kill switch we need some generated
interop bindings. This patch moves both projects to a solution
folder and includes a small modification to toggle the switch if
required.

In addition, somehow the project GUID had become mismatched in the
solution and project files. This is corrected.

Dependencies for Dap.Bluetooth are yet to be checked during
configure, compile with care.