Dap.Bluetooth: Add Crawler
With a minor stateful concession, this patch attempts to be more
conservative in its use of the FTP connection. The previous method
involved ensuring that the working remote folder is always the root at
the start of an invocation.

e.g. if the user wished to transfer an
album, then we'd do two change directory operations, the transfer
and another two directory changes for every file.

Here we introduce a type to hold a small amount of state allowing us
to navigate without worrying where the 'drive head' is.