build: Add quoting and whitespace correction
As reported in BGO#733151, the build system fails after the above
commit. Since m4 uses # as a comment and there is no escaping mechanism
we must use different symbol.

This commit also adds the correct opening backtick for an escaped string
in the mono and bluetoothdap m4 files-after taking a look to see what
the corresponding C# error message was and realising that I had copied
and pasted the error into BluetoothDap.

The symbol is unicode sharp (as in, music): \u266f.

Finally, whitespace in bluetoothdap.m4 is updated in line with the rest
of the build system (i.e. noexpandtab)

Conflicts:
	build/m4/extensions/bluetoothdap.m4