Dap.Bluetooth: Sleep between root requests
The last few patches made it quite easy to trigger a stack overflow
due to the retry being too fast.

When attempting to restore the
connection under these conditions it makes sense to sleep between
recursive calls to allow time for the connection to come back with a
grace time of up to 30 seconds, as originally intended.