Dap.Bluetooth: Crawler, extract Drop to function
This patch extracts the member function, Drop, to a let binding over our
state.

It is used under failure conditions to request that BlueZ
discards the session and allows reconnect in a single
connection scenario (the scenario I test in) where the original
connection is dead, but BlueZ hasn't become aware.