SongKick.GeoLocation: improve notifyUser function
If number of gigs in the user city is a lot Banshee will try
to show many notifications.
What we're doing now is if number of gigs exceeds 3 gigs
then instead of showing all of them, we show one notification
with next text: "SongKick found many gigs near you,
check out all of them in Banshee!"
If we can't find any gigs we also notify user about it.