Dap.Bluetooth: Synchronous sync
In the context of a scheduled connection and synchronisation we need to
know when the sync completes in order to be able to remove the source at
the correct time.

Ideally, DapSync would provide us with a synchronous method. In this
patch, we use Reflection to get at the private method and call it,
falling back to the asynchronous call if we're not successful.