Dap.Bluetooth: Extract Gnome specific component
All AdapterManager really needs is a function to set and unset the
killswitch. We extract this to GnomeRfkill, although both types are
still placed in the same file, and pass an instance's Set member into
AdapterManager at construction time.

The value passed must have signature (bool -> bool), at the moment the
return is unchecked but in future will be used to present a Hyena
Warnf message.