Dap.Bluetooth: Ensure icons are all the same size
Gnome 3.12 brought a size difference between the default icon size for
buttons compared to 3.10. This patch states that LargeToolbar is correct
to use throughout the DeviceWidget, ensuring that buttons line up as if
in a grid.

We leave the MediaControlWidget with smaller buttons, as it is a
sub panel of the DeviceWidget.