Dap.Bluetooth: Retire MediaControlWidget
After updating to Gnome 3.12, I wanted to use a Popover for the media
controls. Unfortunately they aren't available in GTK# yet so this patch
reduces the wasted space by abusing a GtkMenu activated by a toggle
button.

While requiring a little more work to make it look good, as a first
refactoring step this gives us a checkpoint we can return to.